+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/BulkPoolQCIT.java
@@ -38,7 +38,7 @@ public class BulkPoolQCIT extends AbstractIT {
 
   @Before
   public void setup() {
-    loginAdmin();
+    login();
   }
 
   private BulkQCPage getEditPage(List<Long> ids) {