+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/LibraryPageIT.java
@@ -33,7 +33,7 @@ public class LibraryPageIT extends AbstractIT {
 
   @Before
   public void setup() {
-    loginAdmin();
+    login();
   }
 
   @Test