+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/RunPageIT.java
@@ -36,7 +36,7 @@ public class RunPageIT extends AbstractIT {
 
   @Before
   public void setup() {
-    loginAdmin();
+    login();
   }
 
   @Test