+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/ServiceRecordPageIT.java
@@ -23,7 +23,7 @@ public class ServiceRecordPageIT extends AbstractIT {
 
   @Before
   public void setup() {
-    loginAdmin();
+    login();
   }
 
   @Test