+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/StudyPageIT.java
@@ -16,7 +16,7 @@ public class StudyPageIT extends AbstractIT {
 
   @Before
   public void setup() {
-    loginAdmin();
+    login();
   }
 
   private StudyPage getStudyPage(Long studyId, Long projectId) {